requirements.txt
pytest==6.2.5
argparse==1.4.0
pyyaml==6.0